Add verfication key to secret
Without this, you can just manipulate the result by having many secrets
and then picking the one later on, as the bot would still show that you
entrusted it with the secret before people started guessing (for
example).

Now, you can publish the verification key before people start guessing
and that guarantees that you can't pull a bait-and-switch.