Add module meta-data
This adds descriptions to all the modules, which is printed in the
third phase of the /help command.

It also allows modules to be marked as hidden, which means they will be
skipped in the third phase of the /help command and are shown in a new
fourth phase ("hidden modules") instead.